requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
cssselect2==0.4.1
idna==3.3
lxml==4.7.1
Pillow==9.0.0
pony==0.7.14
psycopg2==2.9.3
reportlab==3.6.5
requests==2.27.1
svglib==1.1.0
tinycss2==1.1.1
urllib3==1.26.8
vk-api==11.9.7
webencodings==0.5.1
